requirements.txt
absl-py==1.0.0
alabaster==0.7.8
ale-py==0.7.3
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
atari-py==0.2.9
attrs==21.4.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
bleach==4.1.0
blinker==1.4
box2d-py==2.3.5
Brlapi==0.7.0
cachetools==5.2.0
carla==0.9.13
catkin-pkg-modules==0.4.24
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.24
d4rl @ file:///home/disc/h.bonnavaud/Bureau/these/Papiers/MFTL/src/environments/d4rl
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dm-control @ git+https://github.com/deepmind/dm_control@41d0c7383153f9ca6c12f8e865ef5e73a98759bd
dm-env==1.5
dm-tree==0.1.7
dnspython==1.16.0
docutils==0.16
dotmap==1.3.30
duplicity==0.8.12.0
empy==3.3.2
entrypoints==0.3
fasteners==0.16.3
filelock==3.4.2
flappy-bird-gym==0.3.0
flatbuffers==1.12
flow==0.0.1
fonttools==4.33.3
future==0.18.2
gast==0.4.0
genmsg==0.5.14
genpy==0.6.9
glfw==2.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
graphviz==0.19.1
grpcio==1.48.0
gym==0.18.3
h5py==3.7.0
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.7
imagesize==1.2.0
importlib-metadata==4.11.2
importlib-resources==5.4.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
labmaze==1.0.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.1
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mujoco==2.2.0
mujoco-py==2.1.2.14
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.8.3
notebook==6.4.6
numpy==1.23.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==20.3
pandas==1.4.4
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.1
proglog==0.1.10
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
ptyprocess==0.7.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pybullet==3.2.4
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pygame==2.0.3
PyGetWindow==0.0.9
pyglet==1.5.15
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
PyOpenGL==3.1.6
pyparsing==2.4.6
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.0
PyScreeze==0.1.28
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-xlib==0.15
pytube==12.1.0
pytweening==1.0.4
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
ratelimit==2.2.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
rospkg==1.4.0
rospkg-modules==1.4.0
rsa==4.9
scipy==1.7.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinx-autodoc-typehints==1.17.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.10
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==3.5.2
wrapt==1.14.1
xkit==0.0.0
zipp==3.7.0
